web: Rename config.ts to default.ts, made config/index and export config as .Config
