+++ web/packages/core/src/internal/player/inner.tsx
@@ -7,7 +7,8 @@ import {
     UnmuteOverlay,
     URLLoadOptions,
     WindowMode,
-} from "../../public/config/load-options";
+    DEFAULT_CONFIG,
+} from "../../public/config";
 import type { MovieMetadata } from "../../movie-metadata";
 import { ruffleShadowTemplate } from "../ui/shadow-template";
 import { text, textAsParagraphs } from "../i18n";
@@ -25,7 +26,6 @@ import { showPanicScreen } from "../ui/panic";
 import { createRuffleBuilder } from "../../load-ruffle";
 import { lookupElement } from "../register-element";
 import { configureBuilder } from "../builder";
-import { DEFAULT_CONFIG } from "../../public/config/config";
 
 const DIMENSION_REGEX = /^\s*(\d+(\.\d+)?(%)?)/;
