+++ web/packages/core/src/internal/player/ruffle-player-element.tsx
@@ -1,7 +1,4 @@
-import type {
-    DataLoadOptions,
-    URLLoadOptions,
-} from "../../public/config/load-options";
+import type { DataLoadOptions, URLLoadOptions } from "../../public/config";
 import type { MovieMetadata } from "../../movie-metadata";
 import { InnerPlayer, ReadyState } from "./inner";
 import { Player } from "../../public/player";
