+++ web/packages/core/src/public-api.ts
@@ -1,10 +1,7 @@
 import { Version } from "./version";
 import { VersionRange } from "./version-range";
 import { SourceAPI } from "./source-api";
-import type {
-    DataLoadOptions,
-    URLLoadOptions,
-} from "./public/config/load-options";
+import type { DataLoadOptions, URLLoadOptions } from "./public/config";
 
 declare global {
     interface Window {
