+++ web/packages/core/src/public-path.ts
@@ -1,4 +1,4 @@
-import { BaseLoadOptions } from "./public/config/load-options";
+import { BaseLoadOptions } from "./public/config";
 import { currentScriptURL, isExtension } from "./current-script";
 
 /**
