+++ web/packages/core/src/public/config/default.ts
