+++ web/packages/core/src/public/config/index.ts
@@ -1,0 +1,2 @@
+export * from "./default";
+export * from "./load-options";
