+++ web/packages/demo/src/main.tsx
@@ -4,13 +4,7 @@ import "./common.css";
 import "./lato.css";
 import "./index.css";
 import { App } from "./App.tsx";
-import {
-    AutoPlay,
-    Letterbox,
-    LogLevel,
-    installRuffle,
-    UnmuteOverlay,
-} from "ruffle-core";
+import { Config, installRuffle } from "ruffle-core";
 
 installRuffle("local");
 
@@ -18,10 +12,10 @@ ReactDOM.createRoot(document.getElementById("root")!).render(
     <React.StrictMode>
         <App
             ruffleBaseConfig={{
-                autoplay: AutoPlay.On,
-                unmuteOverlay: UnmuteOverlay.Hidden,
-                logLevel: LogLevel.Warn,
-                letterbox: Letterbox.On,
+                autoplay: Config.AutoPlay.On,
+                unmuteOverlay: Config.UnmuteOverlay.Hidden,
+                logLevel: Config.LogLevel.Warn,
+                letterbox: Config.Letterbox.On,
                 forceScale: true,
                 forceAlign: true,
             }}
