+++ web/packages/extension/src/utils.ts
@@ -1,8 +1,8 @@
 import type { Options } from "./common";
-import { DEFAULT_CONFIG as CORE_DEFAULT_CONFIG } from "ruffle-core";
+import { Config } from "ruffle-core";
 
 const DEFAULT_OPTIONS: Required<Options> = {
-    ...CORE_DEFAULT_CONFIG,
+    ...Config.DEFAULT_CONFIG,
     ruffleEnable: true,
     ignoreOptout: false,
     autostart: false,
